# src/main/java/swagger/api/MoviesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package swagger.api;

import swagger.model.MovieModelGet;
import swagger.model.MovieModelPut;
import swagger.model.ResponseMessage;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;

@Api(value = "movies", description = "the movies API")
public interface MoviesApi {

    @ApiOperation(value = "information about a movie", nickname = "createMovie", notes = "Details of a movie", response = ResponseMessage.class, tags={ "Movie information", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Success", response = ResponseMessage.class),
        @ApiResponse(code = 400, message = "bad request", response = ResponseMessage.class),
        @ApiResponse(code = 500, message = "internal server error", response = ResponseMessage.class),
        @ApiResponse(code = 200, message = "default response message", response = ResponseMessage.class) })
    @RequestMapping(value = "/movies",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<ResponseMessage> createMovie(@ApiParam(value = "body of the movies" ,required=true )  @Valid @RequestBody MovieModelPut body);


    @ApiOperation(value = "information about a movie.", nickname = "outputMovie", notes = "Details of a movie", response = MovieModelGet.class, tags={ "Movie information", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Success", response = MovieModelGet.class) })
    @RequestMapping(value = "/movies",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<MovieModelGet> outputMovie();

}