# src/main/java/swagger/api/RatingsApiController.java
package swagger.api;

import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.annotations.ApiParam;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestBody;
import swagger.model.RatingModel;
import swagger.model.ResponseMessage;
import swagger.service.RatingService;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import java.io.IOException;

@Controller
public class RatingsApiController implements RatingsApi {

    private static final Logger log = LoggerFactory.getLogger(RatingsApiController.class);

    private final ObjectMapper objectMapper;

    private final HttpServletRequest request;

    @Autowired
    RatingService ratingService;

    @Autowired
    public RatingsApiController(ObjectMapper objectMapper, HttpServletRequest request) {
        this.objectMapper = objectMapper;
        this.request = request;
    }

    public ResponseEntity<ResponseMessage> createRating(@ApiParam(value = "body of the movies" ,required=true )  @Valid @RequestBody RatingModel body) {
        String accept = request.getHeader("Accept");
        if (accept != null && accept.contains("application/json")) {
            try {
                ratingService.saveRating(body);
                return new ResponseEntity<ResponseMessage>(objectMapper.readValue("{  \"code\" : 201,  \"message\" : \"Success\"}", ResponseMessage.class), HttpStatus.CREATED);
            } catch (IOException e) {
                log.error("Couldn't serialize response for content type application/json", e);
                return new ResponseEntity<ResponseMessage>(HttpStatus.INTERNAL_SERVER_ERROR);
            }
        }

        return new ResponseEntity<ResponseMessage>(HttpStatus.NOT_IMPLEMENTED);
    }

}